Split up plugin_arena.cc
