+++ firmware/src/memory/plugin_arena.cc
@@ -1,11 +1,8 @@
 #include "plugin_arena.hh"
-#include "system/abort_rescue.hh"
 #include "system/safe_log.hh"
 #include "tlsf/mm_tlsf.h"
 #include <atomic>
-#include <cstdlib>
 #include <cstring>
-#include <new>
 #include <sys/lock.h>
 
 // Defined by the linker: end of the A7 heap region [_sheap, _eheap), which
@@ -45,10 +42,14 @@ inline constexpr size_t MallocAlign = 8;
 // ---- Dynamic boundary ------------------------------------------------------
 // The arena claims slabs downward from _eheap, so the arena region is always
 // the contiguous range [arena_floor, _eheap) and contains() stays a two-
-// comparison check. The firmware heap may grow up to arena_floor (sbrk asks
-// via mm_arena_allow_brk below); slab claims must leave FwHeadroom above the
-// firmware's high-water break. All boundary changes happen under the arena
-// lock; contains() reads the floor lock-free (acquire pairs with the release
+// comparison check.
+// The firmware heap may grow up to arena_floor (sbrk asks via mm_arena_allow_brk)
+//
+// slab claims must leave FwHeadroom above the firmware's high-water break.
+//
+// All boundary changes happen under the arena lock.
+//
+// contains() reads the floor lock-free (acquire pairs with the release
 // store when the floor moves).
 
 // Sentinel: arena empty. Makes contains() false with no extra branch.
@@ -345,232 +346,5 @@ uintptr_t mm_arena_floor(void) {
 }
 
 } // extern "C"
-
-} // namespace MetaModule::PluginArena
-
-using namespace MetaModule;
-
-// ---- free/realloc dispatchers (--wrap) -------------------------------------
-// Ownership legitimately crosses the firmware/plugin boundary in both
-// directions (e.g. a std::string handed across), so every free/realloc must
-// dispatch by the pointer's address. The firmware links with --wrap=free etc.
-// (see src/CMakeLists.txt), which redirects every firmware call site --
-// including newlib-internal ones -- here. Plugin imports of free/_free_r are
-// redirected here as well (allocator_redirect below): --wrap only rewrites
-// link-time references, so a name lookup of "free" in the ELF would find the
-// real newlib function, not the wrapper.
-
-extern "C" {
-
-void __real_free(void *);
-void __real__free_r(struct _reent *, void *);
-void *__real_realloc(void *, size_t);
-void *__real__realloc_r(struct _reent *, void *, size_t);
-
-void __wrap_free(void *p) {
-	if (PluginArena::contains(p))
-		PluginArena::free(p);
-	else
-		__real_free(p);
-}
-
-void __wrap__free_r(struct _reent *r, void *p) {
-	if (PluginArena::contains(p))
-		PluginArena::free(p);
-	else
-		__real__free_r(r, p);
-}
-
-void *__wrap_realloc(void *p, size_t size) {
-	if (PluginArena::contains(p))
-		return PluginArena::realloc(p, size);
-	return __real_realloc(p, size);
-}
-
-void *__wrap__realloc_r(struct _reent *r, void *p, size_t size) {
-	if (PluginArena::contains(p))
-		return PluginArena::realloc(p, size);
-	return __real__realloc_r(r, p, size);
-}
-
-} // extern "C"
-
-// ---- Plugin allocator wrappers ---------------------------------------------
-// Plugins statically link their own operator new/delete/std containers, which
-// call the C allocator functions imported from firmware. The dynloader
-// rebinds those imports (via allocator_redirect below) to these wrappers so
-// that everything plugin code allocates lands in the arena.
-//
-// A failed arena allocation returns nullptr, same as malloc: the plugin's own
-// operator new then throws its own bad_alloc, which the plugin can catch and
-// recover from; if it doesn't, its terminate handler calls the imported
-// abort() -> firmware _kill/_exit (the stage-2 recovery hook).
-
-namespace
-{
-
-void *plugin_realloc(void *old, size_t size) {
-	// A firmware-heap block stays on the firmware heap (newlib owns it); a
-	// null or arena pointer (re)allocates in the arena -- unlike __wrap_realloc,
-	// the caller here is known to be plugin code
-	if (old && !PluginArena::contains(old))
-		return __real_realloc(old, size);
-	return PluginArena::realloc(old, size);
-}
-
-} // namespace
-
-extern "C" {
-
-void *mm_plugin_malloc(size_t size) {
-	return PluginArena::alloc(size);
-}
-
-void *mm_plugin_calloc(size_t nmemb, size_t size) {
-	auto total = nmemb * size;
-	auto *p = PluginArena::alloc(total);
-	if (p)
-		memset(p, 0, total);
-	return p;
-}
-
-void *mm_plugin_realloc(void *old, size_t size) {
-	return plugin_realloc(old, size);
-}
-
-void *mm_plugin_memalign(size_t align, size_t size) {
-	return PluginArena::alloc_aligned(align, size);
-}
-
-int mm_plugin_posix_memalign(void **memptr, size_t align, size_t size) {
-	auto *p = PluginArena::alloc_aligned(align, size);
-	if (!p)
-		return 12; // ENOMEM
-	*memptr = p;
-	return 0;
-}
-
-void *mm_plugin_aligned_alloc(size_t align, size_t size) {
-	return PluginArena::alloc_aligned(align, size);
-}
-
-// newlib reentrant entry points: same behavior, reent unused
-void *mm_plugin_malloc_r(struct _reent *, size_t size) {
-	return PluginArena::alloc(size);
-}
-
-void *mm_plugin_realloc_r(struct _reent *, void *old, size_t size) {
-	return plugin_realloc(old, size);
-}
-
-void *mm_plugin_calloc_r(struct _reent *, size_t nmemb, size_t size) {
-	return mm_plugin_calloc(nmemb, size);
-}
-
-void *mm_plugin_memalign_r(struct _reent *, size_t align, size_t size) {
-	return mm_plugin_memalign(align, size);
-}
-
-// C++ operator new/delete, for plugins that import them instead of statically
-// linking libstdc++'s. The throwing news must not return nullptr; the thrown
-// bad_alloc is the firmware's, and (until cross-boundary unwinding exists)
-// will terminate if no handler is reached before plugin frames.
-void *mm_plugin_op_new(unsigned size) {
-	auto *p = PluginArena::alloc(size);
-	if (!p)
-		throw std::bad_alloc{};
-	return p;
-}
-
-void *mm_plugin_op_new_nothrow(unsigned size, void const *) {
-	return PluginArena::alloc(size);
-}
-
-void *mm_plugin_op_new_aligned(unsigned size, unsigned align) {
-	auto *p = PluginArena::alloc_aligned(align, size);
-	if (!p)
-		throw std::bad_alloc{};
-	return p;
-}
-
-// A plugin that hits an unrecoverable error calls its imported abort() --
-// most commonly via its own std::terminate after an uncaught bad_alloc (its
-// verbose-terminate message has already been printed by the time we get
-// here). Plugin exceptions cannot unwind into firmware frames, so this hook
-// is the only chance to regain control: longjmp back to the rescue scope
-// armed around the call into plugin code (module creation, plugin init).
-// With no scope armed (e.g. death in the audio context), fall through to a
-// real abort and halt.
-void mm_plugin_abort(void) {
-	{
-		SafeLog log;
-		log.str("[abort] plugin called abort(); arena used=");
-		log.u64(PluginArena::used_bytes());
-		log.str(" of ");
-		log.u64(PluginArena::claimed_bytes());
-		log.flush();
-	}
-	MetaModule::abort_rescue_try("plugin abort");
-	::abort();
-}
-
-void mm_plugin_op_delete(void *p) {
-	__wrap_free(p);
-}
-
-void mm_plugin_op_delete_sized(void *p, unsigned) {
-	__wrap_free(p);
-}
-
-void mm_plugin_op_delete_aligned(void *p, unsigned) {
-	__wrap_free(p);
-}
-
-} // extern "C"
-
-namespace MetaModule::PluginArena
-{
-
-void *allocator_redirect(std::string_view name) {
-	// clang-format off
-	if (name == "malloc")          return reinterpret_cast<void *>(mm_plugin_malloc);
-	if (name == "calloc")          return reinterpret_cast<void *>(mm_plugin_calloc);
-	if (name == "realloc")         return reinterpret_cast<void *>(mm_plugin_realloc);
-	if (name == "memalign")        return reinterpret_cast<void *>(mm_plugin_memalign);
-	if (name == "posix_memalign")  return reinterpret_cast<void *>(mm_plugin_posix_memalign);
-	if (name == "aligned_alloc")   return reinterpret_cast<void *>(mm_plugin_aligned_alloc);
-	if (name == "_malloc_r")       return reinterpret_cast<void *>(mm_plugin_malloc_r);
-	if (name == "_realloc_r")      return reinterpret_cast<void *>(mm_plugin_realloc_r);
-	if (name == "_calloc_r")       return reinterpret_cast<void *>(mm_plugin_calloc_r);
-	if (name == "_memalign_r")     return reinterpret_cast<void *>(mm_plugin_memalign_r);
-
-	// De-allocation goes to the address-dispatching wrappers: a name lookup
-	// of "free" in the ELF finds real newlib free (--wrap doesn't rename), so
-	// plugins would otherwise free arena pointers into the newlib heap
-	if (name == "free")            return reinterpret_cast<void *>(__wrap_free);
-	if (name == "_free_r")         return reinterpret_cast<void *>(__wrap__free_r);
-
-	// Plugin death (uncaught exception -> terminate -> abort) routes to the
-	// rescue hook instead of halting
-	if (name == "abort")           return reinterpret_cast<void *>(mm_plugin_abort);
-
-	// operator new(unsigned), new[], nothrow and aligned variants (arm32 mangling)
-	if (name == "_Znwj")                     return reinterpret_cast<void *>(mm_plugin_op_new);
-	if (name == "_Znaj")                     return reinterpret_cast<void *>(mm_plugin_op_new);
-	if (name == "_ZnwjRKSt9nothrow_t")       return reinterpret_cast<void *>(mm_plugin_op_new_nothrow);
-	if (name == "_ZnajRKSt9nothrow_t")       return reinterpret_cast<void *>(mm_plugin_op_new_nothrow);
-	if (name == "_ZnwjSt11align_val_t")      return reinterpret_cast<void *>(mm_plugin_op_new_aligned);
-	if (name == "_ZnajSt11align_val_t")      return reinterpret_cast<void *>(mm_plugin_op_new_aligned);
-
-	// operator delete(void*), delete[], sized and aligned variants
-	if (name == "_ZdlPv")                    return reinterpret_cast<void *>(mm_plugin_op_delete);
-	if (name == "_ZdaPv")                    return reinterpret_cast<void *>(mm_plugin_op_delete);
-	if (name == "_ZdlPvj")                   return reinterpret_cast<void *>(mm_plugin_op_delete_sized);
-	if (name == "_ZdaPvj")                   return reinterpret_cast<void *>(mm_plugin_op_delete_sized);
-	if (name == "_ZdlPvSt11align_val_t")     return reinterpret_cast<void *>(mm_plugin_op_delete_aligned);
-	if (name == "_ZdaPvSt11align_val_t")     return reinterpret_cast<void *>(mm_plugin_op_delete_aligned);
-	// clang-format on
-	return nullptr;
-}
 
 } // namespace MetaModule::PluginArena
